requirements-test.txt
pytest>=7.4
pytest-mock>=3.12
apache-airflow==2.9.0
pyyaml>=6.0
croniter>=2.0
boto3>=1.34
